make SIMFracture(Qstatic|Miehe) a private template
